creating an oop structure including tests on the blockchain and miningchain
